Make encryption optional
Encryption is now optional and disabled by default, unless the `secret` configuration option is set. This ensures existing installations keep working as expected.

The reason for disabling encryption by default is in order for us to be able to use utility commands to setup encryption, for example, generating keys. This would not be possible if encryption would be enabled by default, in which case configuration validation would fail when trying to run the command.